requirements.txt
Flask==1.1.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Werkzeug==1.0.1
regex==2020.11.13
allennlp==1.3.0 
tokenizers==0.9.4
scikit-learn==0.23.2
pandas==1.1.4
numpy==1.18.5
nltk==3.5
bert-extractive-summarizer==0.5.2
lxml==4.6.2 
python-docx==0.8.10
PyPDF2==1.26.0
striprtf==0.0.12
XlsxWriter==1.3.7 
python-pptx==0.6.18
